+++ apps/client/src/app/directives/file-drop/file-drop.directive.ts
@@ -1,4 +1,4 @@
-import { Directive, HostListener, Output, EventEmitter } from '@angular/core';
+import { Directive, EventEmitter, HostListener, Output } from '@angular/core';
 
 @Directive({
   selector: '[gfFileDrop]'