+++ apps/client/src/app/pages/features/features-page.component.ts
@@ -15,6 +15,7 @@ export class FeaturesPageComponent implements OnDestroy {
   public hasPermissionForSubscription: boolean;
   public info: InfoItem;
   public routerLinkRegister = ['/' + $localize`register`];
+  public routerLinkResources = ['/' + $localize`resources`];
   public user: User;
 
   private unsubscribeSubject = new Subject<void>();