+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Supported enter key press to submit the form of the create or update account dialog
 - Added the version to the admin control panel
 - Added pagination parameters (`skip`, `take`) to the endpoint `GET api/v1/order`