+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added the application version to the endpoint `GET api/v1/admin`
 
+### Fixed
+
+- Fixed the style of the active features page in the navigation on desktop
+
 ## 2.8.0 - 2023-10-03
 
 ### Added