+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added support for notes in the activities import
 - Added the application version to the endpoint `GET api/v1/admin`
 
 ### Fixed